Folder refactoring
