+++ app/src/main/java/com/example/vinilos/models/Comment.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.model
+package com.example.vinilos.models
 
 data class Comment (
     var id: Int?,
