+++ app/src/main/java/com/example/vinilos/models/Perfomer.kt
@@ -1,6 +1,4 @@
-package com.example.vinilos.model
-
-import java.util.*
+package com.example.vinilos.models
 
 data class Perfomer (
