+++ app/src/main/java/com/example/vinilos/providers/NetworkServiceAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.provider
+package com.example.vinilos.providers
 
 import android.content.Context
 import com.android.volley.Request
@@ -8,7 +8,7 @@ import com.android.volley.VolleyError
 import com.android.volley.toolbox.JsonObjectRequest
 import com.android.volley.toolbox.StringRequest
 import com.android.volley.toolbox.Volley
-import com.example.vinilos.model.Album
+import com.example.vinilos.models.Album
 import com.fasterxml.jackson.module.kotlin.*
 import org.json.JSONObject
