+++ app/src/main/java/com/example/vinilos/repositories/AlbumRepository.kt
@@ -1,9 +1,9 @@
-package com.example.vinilos.repository
+package com.example.vinilos.repositories
 
 import android.app.Application
 import com.android.volley.VolleyError
-import com.example.vinilos.model.Album
-import com.example.vinilos.provider.NetworkServiceAdapter
+import com.example.vinilos.models.Album
+import com.example.vinilos.providers.NetworkServiceAdapter
 
 class AlbumRepository (val application: Application){
     fun refreshData(callback: (List<Album>)->Unit, onError: (VolleyError)->Unit) {
