+++ app/src/main/java/com/example/vinilos/ui/fragments/ArtistsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos.ui.collectors
+package com.example.vinilos.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,12 +8,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.example.vinilos.databinding.FragmentCollectorsBinding
+import com.example.vinilos.databinding.FragmentArtistsBinding
+import com.example.vinilos.ui.viewmodels.ArtistsViewModel
 
-class CollectorsFragment : Fragment() {
+class ArtistsFragment : Fragment() {
 
-    private lateinit var collectorsViewModel: CollectorsViewModel
-    private var _binding: FragmentCollectorsBinding? = null
+    private lateinit var artistsViewModel: ArtistsViewModel
+    private var _binding: FragmentArtistsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,14 +25,14 @@ class CollectorsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        collectorsViewModel =
-            ViewModelProvider(this).get(CollectorsViewModel::class.java)
+        artistsViewModel =
+            ViewModelProvider(this).get(ArtistsViewModel::class.java)
 
-        _binding = FragmentCollectorsBinding.inflate(inflater, container, false)
+        _binding = FragmentArtistsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textCollectors
-        collectorsViewModel.text.observe(viewLifecycleOwner, Observer {
+        val textView: TextView = binding.textArtists
+        artistsViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
