+++ app/src/main/java/com/example/vinilos/ui/viewmodels/ArtistsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.vinilos.ui.collectors
+package com.example.vinilos.ui.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class CollectorsViewModel : ViewModel() {
+class ArtistsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is collectors Fragment"
+        value = "This is artists Fragment"
     }
     val text: LiveData<String> = _text
 }
