+++ app/src/main/res/layout/fragment_albums.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.albums.AlbumsFragment">
+    tools:context=".ui.fragments.AlbumsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/fragment_album"
