+++ app/src/main/res/layout/fragment_artists.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.artists.ArtistsFragment">
+    tools:context=".ui.fragments.ArtistsFragment">
 
     <TextView
         android:id="@+id/text_artists"
