+++ app/src/main/res/layout/fragment_guest.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.guest.GuestFragment">
+    tools:context=".ui.fragments.GuestFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
