+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/nav_guest"
-        android:name="com.example.vinilos.ui.guest.GuestFragment"
+        android:name="com.example.vinilos.ui.fragments.GuestFragment"
         android:label="@string/menu_guest"
         tools:layout="@layout/fragment_guest" />
 
     <fragment
         android:id="@+id/nav_artists"
-        android:name="com.example.vinilos.ui.artists.ArtistsFragment"
+        android:name="com.example.vinilos.ui.fragments.ArtistsFragment"
         android:label="@string/menu_artists"
         tools:layout="@layout/fragment_artists" />
 
     <fragment
         android:id="@+id/nav_albums"
-        android:name="com.example.vinilos.ui.albums.AlbumsFragment"
+        android:name="com.example.vinilos.ui.fragments.AlbumsFragment"
         android:label="@string/menu_albums"
         tools:layout="@layout/fragment_albums" />
 
     <fragment
         android:id="@+id/nav_collectors"
-        android:name="com.example.vinilos.ui.collectors.CollectorsFragment"
+        android:name="com.example.vinilos.ui.fragments.CollectorsFragment"
         android:label="@string/menu_collectors"
         tools:layout="@layout/fragment_collectors" />
 </navigation>
